requirements.txt
pytest==8.3.2
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
Faker==28.4.1
allure-pytest==2.13.5
jsonschema~=4.25.1
dotenv~=0.9.9